+++ Sourcery/CodeGenerated/Equality.generated.swift
@@ -1,11 +1,11 @@
-// Generated using Sourcery 0.5.0 — https://github.com/krzysztofzablocki/Sourcery
+// Generated using Sourcery 0.5.1 — https://github.com/krzysztofzablocki/Sourcery
 // DO NOT EDIT
 
 extension Enum {
     override func isEqual(_ object: Any?) -> Bool {
         guard let rhs = object as? Enum else { return false }
         if self.cases != rhs.cases { return false }
-        if self.rawType != rhs.rawType { return false }
+        if self.rawTypeName != rhs.rawTypeName { return false }
 
         return super.isEqual(rhs)
     }