+++ Sourcery/Models/Annotations.swift
@@ -1,0 +1,5 @@
+import Foundation
+
+protocol Annotated {
+    var annotations: [String: NSObject] { get }
+}