+++ SourceryTests/Generating/StencilTemplateSpec.swift
@@ -19,16 +19,22 @@ class StencilTemplateSpec: QuickSpec {
             it("checks for string in name") {
                 expect(generate("{{ \"FooBar\" | contains:\"oo\" }}")).to(equal("true"))
                 expect(generate("{{ \"FooBar\" | contains:\"xx\" }}")).to(equal("false"))
+                expect(generate("{{ \"FooBar\" | !contains:\"oo\" }}")).to(equal("false"))
+                expect(generate("{{ \"FooBar\" | !contains:\"xx\" }}")).to(equal("true"))
             }
 
             it("checks for string in prefix") {
                 expect(generate("{{ \"FooBar\" | hasPrefix:\"Foo\" }}")).to(equal("true"))
                 expect(generate("{{ \"FooBar\" | hasPrefix:\"Bar\" }}")).to(equal("false"))
+                expect(generate("{{ \"FooBar\" | !hasPrefix:\"Foo\" }}")).to(equal("false"))
+                expect(generate("{{ \"FooBar\" | !hasPrefix:\"Bar\" }}")).to(equal("true"))
             }
 
             it("checks for string in suffix") {
                 expect(generate("{{ \"FooBar\" | hasSuffix:\"Bar\" }}")).to(equal("true"))
                 expect(generate("{{ \"FooBar\" | hasSuffix:\"Foo\" }}")).to(equal("false"))
+                expect(generate("{{ \"FooBar\" | !hasSuffix:\"Bar\" }}")).to(equal("false"))
+                expect(generate("{{ \"FooBar\" | !hasSuffix:\"Foo\" }}")).to(equal("true"))
             }
 
         }