+++ SourceryTests/GeneratorSpec.swift
@@ -10,8 +10,8 @@ class GeneratorSpec: QuickSpec {
         describe("Generator") {
 
             let fooType = Class(name: "Foo", variables: [Variable(name: "intValue", typeName: "Int")], inheritedTypes: ["NSObject", "Decodable", "AlternativeProtocol"])
-            let fooSubclassType = Class(name: "FooSubclass", inheritedTypes: ["Foo", "ProtocolBasedOnKnownProtocol"])
-            let barType = Struct(name: "Bar", inheritedTypes: ["KnownProtocol", "Decodable"])
+            let fooSubclassType = Class(name: "FooSubclass", inheritedTypes: ["Foo", "ProtocolBasedOnKnownProtocol"], annotations: ["foo": NSNumber(value: 2)])
+            let barType = Struct(name: "Bar", inheritedTypes: ["KnownProtocol", "Decodable"], annotations: ["bar": NSNumber(value: true)])
 
             let complexType = Struct(name: "Complex", accessLevel: .public, isExtension: false, variables: [])
             let fooVar = Variable(name: "foo", typeName: "Foo", accessLevel: (read: .public, write: .public), isComputed: false)
@@ -42,7 +42,7 @@ class GeneratorSpec: QuickSpec {
                             Variable(name: "foo", typeName: "Int", accessLevel: (read: .public, write: .public), isComputed: false)
                             ])
                         ]),
-                    Enum(name: "FooOptions", accessLevel: .public, inheritedTypes: ["Foo", "KnownProtocol"], rawType: "Foo", cases: [Enum.Case(name: "fooA"), Enum.Case(name: "fooB")]),
+                    Enum(name: "FooOptions", accessLevel: .public, inheritedTypes: ["Foo", "KnownProtocol"], rawTypeName: "Foo", cases: [Enum.Case(name: "fooA"), Enum.Case(name: "fooB")]),
                     Type(name: "NSObject", accessLevel: .none, isExtension: true, inheritedTypes: ["KnownProtocol"]),
                     Class(name: "ProjectClass", accessLevel: .none),
                     Class(name: "ProjectFooSubclass", inheritedTypes: ["FooSubclass"]),
@@ -54,7 +54,7 @@ class GeneratorSpec: QuickSpec {
             let arguments: [String: NSObject] = ["some": "value" as NSString, "number": NSNumber(value: Float(4))]
 
             func generate(_ template: String) -> String {
-                let types = ParserComposer().uniqueTypes((types, []))
+                let types = Composer().uniqueTypes((types, []))
 
                 return (try? Generator.generate(types,
                         template: StencilTemplate(templateString: template),
@@ -121,6 +121,11 @@ class GeneratorSpec: QuickSpec {
 
                 it("counts all variables including implements, inherits") {
                     expect(generate("{{ type.ProjectFooSubclass.allVariables.count }}")).to(equal("2"))
+                }
+
+                it("can use annotations filter") {
+                    expect(generate("{% for type in types.all|annotated:\"bar\" %}{{ type.name }}{% endfor %}")).to(equal("Bar"))
+                    expect(generate("{% for type in types.all|annotated:\"foo = 2\" %}{{ type.name }}{% endfor %}")).to(equal("FooSubclass"))
                 }
 
                 it("can use filter on variables") {