+++ SourceryTests/Models/VariableSpec.swift
@@ -23,47 +23,28 @@ class VariableSpec: QuickSpec {
                 expect(sut?.writeAccess == AccessLevel.internal.rawValue).to(beTrue())
             }
 
-            it("reports optional false") {
-                expect(sut?.isOptional).to(beFalse())
+            it("can report optional via KVC") {
+                expect(Variable(name: "Foo", typeName: "Int?").value(forKeyPath: "isOptional") as? Bool).to(equal(true))
+                expect(Variable(name: "Foo", typeName: "Int!").value(forKeyPath: "isOptional") as? Bool).to(equal(true))
+                expect(Variable(name: "Foo", typeName: "Int?").value(forKeyPath: "isImplicitlyUnwrappedOptional") as? Bool).to(equal(false))
+                expect(Variable(name: "Foo", typeName: "Int!").value(forKeyPath: "isImplicitlyUnwrappedOptional") as? Bool).to(equal(true))
+
+                expect(Variable(name: "Foo", typeName: "Int?").value(forKeyPath: "unwrappedTypeName") as? String).to(equal("Int"))
             }
 
-            context("given optional type with short syntax") {
-                it("reports optional true") {
-                    expect(Variable(name: "Foo", typeName: "Int?").isOptional).to(beTrue())
-                }
+            it("can report tuple type via KVC") {
+                let tuple = TupleType(name: "(Int, Int)", elements: [])
+                let variable = Variable(name: "Foo", typeName: "(Int, Int)")
+                variable.typeName.tuple = tuple
 
-                it("reports non-optional type for unwrappedTypeName") {
-                    expect(Variable(name: "Foo", typeName: "Int?").unwrappedTypeName).to(equal("Int"))
-                }
-
-                it("can report optional via KVC") {
-                    let variable = Variable(name: "Foo", typeName: "Int?")
-
-                    expect(variable.value(forKeyPath: "isOptional") as? Bool).to(equal(true))
-                }
+                expect(variable.value(forKeyPath: "isTuple") as? Bool).to(equal(true))
             }
 
-            context("given optional type with long generic syntax") {
-                it("reports optional true") {
-                    expect(Variable(name: "Foo", typeName: "Optional<Int>").isOptional).to(beTrue())
-                }
+            it("can report actual type name via KVC") {
+                let variable = Variable(name: "Foo", typeName: "Alias")
+                variable.typeName.actualTypeName = TypeName("Int")
 
-                it("reports non-optional type for unwrappedTypeName") {
-                    expect(Variable(name: "Foo", typeName: "Optional<Int>").unwrappedTypeName).to(equal("Int"))
-                }
-            }
-
-            describe("tuple type") {
-                it("reports tuple correctly") {
-                    expect(TypeName("(Int, Int)").isTuple).to(beTrue())
-                    expect(TypeName("(Int)").isTuple).to(beFalse())
-                    expect(TypeName("Int").isTuple).to(beFalse())
-                    expect(TypeName("(Int) -> (Int)").isTuple).to(beFalse())
-                    expect(TypeName("(Int, Int) -> (Int)").isTuple).to(beFalse())
-                    expect(TypeName("(Int, (Int, Int) -> (Int))").isTuple).to(beTrue())
-                    expect(TypeName("(Int, (Int, Int))").isTuple).to(beTrue())
-                    expect(TypeName("(Int, (Int) -> (Int -> Int))").isTuple).to(beTrue())
-                }
+                expect(variable.value(forKeyPath: "actualTypeName") as? TypeName).to(equal(variable.typeName.actualTypeName))
             }
 
             describe("When testing equality") {